Add println!() support, working in Dolphin
Once #9 gets fixed to work with a real USB Gecko, this will have to do
as a way to communicate with the outside world for debugging purposes.

I quite dislike how the alloc and luma_core crates must be imported
under these names in order for my macro to work, but I couldn’t figure
out a better way which still compiles.