Add a module for IPC with the Starlet running IOS
This lets us open file descriptors and perform actions on them.

MINI uses a different protocol, so this can’t currently be used to talk
with this Starlet firmware, see https://wiibrew.org/wiki/MINI

TODO: Figure out why it only works on Dolphin and not on a real Wii.